requirements.txt
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
idna==3.8
pycparser==2.22
pyOpenSSL==24.2.1
requests==2.32.3
urllib3==2.2.2
